% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{nbLackOfFitKL}
\alias{nbLackOfFitKL}
\title{Calcualte the KL divergence of a negative binomial fit to the mismatch data.}
\usage{
nbLackOfFitKL(res)
}
\arguments{
\item{res}{A TEfit object.}
}
\description{
Calcualte the KL divergence of a negative binomial fit to the mismatch data.
}
\examples{
# Analyze Gypsy family 24 (Nusif)
data(AetLTR)
dat <- subset(AetLTR, GroupID == 24 & !is.na(Chr))
set.seed(1)
res1 <- EstDynamics(dat$Mismatch, dat$UngapedLen, plotFit=TRUE, plotSensitivity=FALSE, pause=FALSE)
nbLackOfFitKL(res1)
}
