% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{print.TEfit}
\alias{print.TEfit}
\alias{print.TEfit2}
\title{Print a TEfit or TEfit2 object}
\usage{
\method{print}{TEfit}(x, ...)

\method{print}{TEfit2}(x, ...)
}
\arguments{
\item{x}{A TEfit or TEfit2 object}

\item{...}{Not used}
}
\description{
Print a TEfit or TEfit2 object
}
